Remove `AttributeLintKind::MalformedDiagnosticFormat` variant
