+++ compiler/rustc_attr_parsing/src/errors.rs
@@ -356,3 +356,20 @@ pub(crate) struct MalFormedDiagnosticAttributeLint {
     #[label("invalid option found here")]
     pub span: Span,
 }
+
+#[derive(Diagnostic)]
+#[diag("positional format arguments are not allowed here")]
+#[help(
+    "only named format arguments with the name of one of the generic types are allowed in this context"
+)]
+pub(crate) struct DisallowedPositionalArgument;
+
+#[derive(Diagnostic)]
+#[diag("format arguments are not allowed here")]
+#[help("consider removing this format argument")]
+pub(crate) struct DisallowedPlaceholder;
+
+#[derive(Diagnostic)]
+#[diag("invalid format specifier")]
+#[help("no format specifier are supported in this position")]
+pub(crate) struct InvalidFormatSpecifier;
