+++ compiler/rustc_lint_defs/src/lib.rs
@@ -656,7 +656,6 @@ pub enum DeprecatedSinceKind {
 pub enum AttributeLintKind {
     UnexpectedCfgName((Symbol, Span), Option<(Symbol, Span)>),
     UnexpectedCfgValue((Symbol, Span), Option<(Symbol, Span)>),
-    MalformedDiagnosticFormat { warning: FormatWarning },
     DiagnosticWrappedParserError { description: String, label: String, span: Span },
     IgnoredDiagnosticOption { option_name: Symbol, first_span: Span, later_span: Span },
     MissingOptionsForDiagnosticAttribute { attribute: &'static str, options: &'static str },
